split up subscriber logic in 3 different types
