+++ src/lib.rs
@@ -2,19 +2,18 @@ mod error;
 mod message;
 mod sequence_message;
 mod subscribe;
-#[cfg(feature = "async")]
-mod subscribe_async;
 
 pub use crate::{
     error::Error,
     message::{Message, DATA_MAX_LEN, SEQUENCE_LEN, TOPIC_MAX_LEN},
     sequence_message::SequenceMessage,
     subscribe::{
-        subscribe_multi, subscribe_multi_blocking, subscribe_single, subscribe_single_blocking,
+        blocking::{subscribe_multi_blocking, subscribe_single_blocking},
+        receiver::{subscribe_multi, subscribe_single},
     },
 };
 
 #[cfg(feature = "async")]
-pub use crate::subscribe_async::{
+pub use crate::subscribe::stream::{
     subscribe_async, subscribe_multi_async, MessageStream, MultiMessageStream,
 };
